fix: crypto in insecure browser context
This PR adds `crypto-browserify` to the dependencies and replaces `crypto` with `crypto-browserify` when bundled in the browser.

In files that require webcrypto we check to see if it's available. If it is not we require the Node.js implementation (which has `crypto` replaced with `crypto-browserify`) and if it is available then we use the webcrypto version (so we get fast crypto).

Shipping `crypto-browserify` adds to the bundle size:

Current gzipped size: 142,824 bytes
New gzipped size: 214,499 bytes

Difference: **+71,675 bytes**

It's not an insignificant addition so we need to decide whether this is worth it.

If not accepted, we need to add checks when libp2p-crypto methods are called and callback with an appropriate error message. JS IPFS will continue to have issues opened with confusion around this otherwise! See https://github.com/ipfs/js-ipfs/issues/963 https://github.com/ipfs/js-ipfs/issues/964 https://github.com/ipfs/js-ipfs/issues/2153

resolves https://github.com/libp2p/js-libp2p-crypto/issues/105

License: MIT
Signed-off-by: Alan Shaw <alan.shaw@protocol.ai>