feat: bring your own web crypto
Changes `webcrypto.js` to check for native web crypto availability and falls
back to using `window.__crypto` if not available.

If the user wants to bring their own Web Crypto API compatible implementation
then they simply need to assign it to `window.__crypto` before they start using
IPFS.

Checks are done in the functions that require web crypto to give the user the
flexibility to assign to `window.__crypto` before OR after they import
`libp2p-crypto`. It also means that users have the ability to use other exported
functions that do not require web crypto without having to worry about sorting
their own implementation.

We use `window.__crypto` because `window.crypto` is a readonly property in
secure context and always readonly in workers.

If `window.crypto` and `window.__cypto` are unavailable then an appropriate
error message is reported to the user with a `ERR_MISSING_WEB_CRYPTO` code.

I've also added documentation to the README.

This is a backwards compatible change.

closes https://github.com/libp2p/js-libp2p-crypto/pull/149
resolves https://github.com/libp2p/js-libp2p-crypto/issues/105
resolves https://github.com/ipfs/js-ipfs/issues/2017
resolves https://github.com/ipfs/js-ipfs/issues/2153

License: MIT
Signed-off-by: Alan Shaw <alan.shaw@protocol.ai>